+++ test-suite/complexity/ring.v
@@ -5,3 +5,4 @@ Require Import ZArith.
 Open Scope Z_scope.
 Goal forall a, a+a+a+a+a+a+a+a+a+a+a+a+a = a*13.
 Timeout 5 Time intro; ring.
+Abort.