+++ test-suite/misc/aux7704.v
@@ -1,4 +1,3 @@
-
 Goal True /\ True.
 Proof.
   split.