+++ toplevel/usage.ml
@@ -23,7 +23,7 @@ let machine_readable_version ret =
 let extra_usage = ref []
 let add_to_usage name text = extra_usage := (name,text) :: !extra_usage
 
-let print_usage_channel co command =
+let print_usage_common co command =
   output_string co command;
   output_string co "Coq options are:\n";
   output_string co
@@ -48,9 +48,6 @@ let print_usage_channel co command =
 \n  -lv f	                 (idem)\
 \n  -load-vernac-object f  load Coq object file f.vo\
 \n  -require path          load Coq library path and import it (Require Import path.)\
-\n  -compile f.v           compile Coq file f.v (implies -batch)\
-\n  -compile-verbose f.v   verbosely compile Coq file f.v (implies -batch)\
-\n  -o f.vo                use f.vo as the output file name\
 \n  -quick                 quickly compile .v files to .vio files (skip proofs)\
 \n  -schedule-vio2vo j f1..fn   run up to j instances of Coq to turn each fi.vio\
 \n                         into fi.vo\
@@ -66,16 +63,15 @@ let print_usage_channel co command =
 \n  -quiet                 unset display of extra information (implies -w \"-all\")\
 \n  -w (w1,..,wn)          configure display of warnings\
 \n  -color (yes|no|auto)   configure color output\
+\n  -emacs                 tells Coq it is executed under Emacs\
 \n\
 \n  -q                     skip loading of rcfile\
 \n  -init-file f           set the rcfile to f\
-\n  -batch                 batch mode (exits just after arguments parsing)\
 \n  -boot                  boot mode (implies -q and -batch)\
 \n  -bt                    print backtraces (requires configure debug flag)\
 \n  -debug                 debug mode (implies -bt)\
 \n  -diffs (on|off|removed) highlight differences between proof steps\
 \n  -stm-debug             STM debug mode (will trace every transaction)\
-\n  -emacs                 tells Coq it is executed under Emacs\
 \n  -noglob                do not dump globalizations\
 \n  -dump-glob f           dump globalizations in file f (to be used by coqdoc)\
 \n  -impredicative-set     set sort Set impredicative\
@@ -101,21 +97,36 @@ let print_usage_channel co command =
 
 (* print the usage on standard error *)
 
-let print_usage = print_usage_channel stderr
-
 let print_usage_coqtop () =
-  print_usage "Usage: coqtop <options>\n\n";
+  print_usage_common stderr "Usage: coqtop <options>\n\n";
+  output_string stderr "\n\
+coqtop specific options:\
+\n\
+\n  -batch                 batch mode (exits just after arguments parsing)\
+\n\
+\nDeprecated options [use coqc instead]:\
+\n\
+\n  -compile f.v           compile Coq file f.v (implies -batch)\
+\n  -compile-verbose f.v   verbosely compile Coq file f.v (implies -batch)\
+\n  -o f.vo                use f.vo as the output file name\
+\n";
   flush stderr ;
   exit 1
 
 let print_usage_coqc () =
-  print_usage "Usage: coqc <options> <Coq options> file...\n\
-\noptions are:\
-\n  -verbose  compile verbosely\
-\n  -image f  specify an alternative executable for Coq\
-\n  -opt      run the native-code version of Coq\
-\n  -byte     run the bytecode version of Coq\
-\n  -t        keep temporary files\n\n";
+  print_usage_common stderr "Usage: coqc <options> <Coq options> file...";
+  output_string stderr "\n\
+coqc specific options:\
+\n\
+\n  -o f.vo                use f.vo as the output file name\
+\n  -verbose               compile and output the input file\
+\n\
+\nDeprecated options:\
+\n\
+\n  -image f               specify an alternative executable for Coq\
+\n  -opt                   run the native-code version of Coq\
+\n  -byte                  run the bytecode version of Coq\
+\n  -t                     keep temporary files\
+\n";
   flush stderr ;
   exit 1
-